+++ example/main.dart
@@ -1,5 +1,5 @@
 import 'dart:convert' as convert;
-import 'package:http/http.dart' as http;
+import 'package:dz_http/http.dart' as http;
 
 void main(List<String> arguments) async {
   // This example uses the Google Books API to search for books about http.