+++ lib/src/client.dart
@@ -39,7 +39,11 @@ abstract class Client {
   /// Sends an HTTP GET request with the given headers to the given URL.
   ///
   /// For more fine-grained control over the request, use [send] instead.
-  Future<Response> get(Uri url, {Map<String, String>? headers});
+  Future<Response> get(
+    Uri url, {
+    Map<String, String>? headers,
+    Map<String, String>? body,
+  });
 
   /// Sends an HTTP POST request with the given headers and body to the given
   /// URL.