+++ test/html/client_test.dart
@@ -4,8 +4,8 @@
 
 @TestOn('browser')
 
-import 'package:http/http.dart' as http;
-import 'package:http/browser_client.dart';
+import 'package:dz_http/http.dart' as http;
+import 'package:dz_http/browser_client.dart';
 import 'package:test/test.dart';
 
 import 'utils.dart';