+++ test/io/client_test.dart
@@ -7,8 +7,8 @@
 import 'dart:convert';
 import 'dart:io';
 
-import 'package:http/http.dart' as http;
-import 'package:http/io_client.dart' as http_io;
+import 'package:dz_http/http.dart' as http;
+import 'package:dz_http/io_client.dart' as http_io;
 import 'package:test/test.dart';
 
 import 'utils.dart';