+++ test/response_test.dart
@@ -4,7 +4,7 @@
 
 import 'dart:async';
 
-import 'package:http/http.dart' as http;
+import 'package:dz_http/http.dart' as http;
 import 'package:pedantic/pedantic.dart';
 import 'package:test/test.dart';